test: move spoofed mast serde tests to miden-mast-package
