+++ crates/assembly/src/tests.rs
@@ -418,188 +418,6 @@ fn static_package_same_digest_procedure_uses_exact_root_metadata() -> Result<(),
     Ok(())
 }
 
-fn read_usize_vint64(bytes: &[u8], offset: &mut usize) -> usize {
-    // This test patches raw bytes in place, so it needs byte offsets that ByteReader::read_usize
-    // does not expose.
-    let first_byte = bytes.get(*offset).copied().expect("out-of-bounds vint64 peek");
-    let length = first_byte.trailing_zeros() as usize + 1;
-
-    if length == 9 {
-        *offset += 1;
-        let end = (*offset).checked_add(8).expect("offset overflow while reading vint64");
-        let chunk: [u8; 8] = bytes[*offset..end].try_into().expect("out-of-bounds vint64");
-        *offset = end;
-        let value = u64::from_le_bytes(chunk);
-        usize::try_from(value).expect("encoded usize does not fit host usize")
-    } else {
-        let end = (*offset).checked_add(length).expect("offset overflow while reading vint64");
-        let mut encoded = [0u8; 8];
-        encoded[..length].copy_from_slice(&bytes[*offset..end]);
-        *offset = end;
-        let value = u64::from_le_bytes(encoded) >> length;
-        usize::try_from(value).expect("encoded usize does not fit host usize")
-    }
-}
-
-fn locate_first_node_hash(bytes: &[u8]) -> (usize, usize) {
-    // Header: magic[4] + flags[1] + version[3]
-    let mut offset = 0usize;
-    offset += 4;
-    offset += 1;
-    offset += 3;
-
-    let internal_node_count = read_usize_vint64(bytes, &mut offset);
-    let external_node_count = read_usize_vint64(bytes, &mut offset);
-    let node_count = internal_node_count
-        .checked_add(external_node_count)
-        .expect("node count overflow");
-
-    // Roots: len (usize) + elements (u32 LE)
-    let roots_len = read_usize_vint64(bytes, &mut offset);
-    offset += roots_len * 4;
-
-    // Basic block data: len (usize) + bytes
-    let bb_len = read_usize_vint64(bytes, &mut offset);
-    offset += bb_len;
-
-    offset += node_count * 8;
-    offset += external_node_count * 32;
-
-    (offset, internal_node_count)
-}
-
-fn build_library_bytes_with_spoofed_first_node_digest(
-    lib: &Package,
-    spoof_seed: &str,
-) -> (Vec<u8>, Word) {
-    use miden_core::serde::Serializable;
-
-    // Serialize the MastForest in stripped form so the byte layout is minimal and stable.
-    let forest = lib.mast_forest().as_ref();
-    let original_digest = forest[MastNodeId::new_unchecked(0)].digest();
-    let mut output_bytes = Vec::new();
-    lib.write_header_into(&mut output_bytes);
-    let forest_offset = output_bytes.len();
-    forest.write_stripped(&mut output_bytes);
-
-    let (node_hashes_start, node_count) = locate_first_node_hash(&output_bytes[forest_offset..]);
-    assert!(node_count > 0, "expected at least one node info entry");
-
-    // Patch node 0 digest in-place.
-    let spoofed_digest = miden_core::utils::hash_string_to_word(spoof_seed);
-    assert_ne!(spoofed_digest, original_digest, "spoofed digest must differ");
-
-    let mut spoofed_digest_bytes = Vec::new();
-    spoofed_digest.write_into(&mut spoofed_digest_bytes);
-    assert_eq!(spoofed_digest_bytes.len(), 32, "Word must serialize to 32 bytes");
-
-    let node0_digest_offset = forest_offset + node_hashes_start;
-    output_bytes[node0_digest_offset..node0_digest_offset + 32]
-        .copy_from_slice(&spoofed_digest_bytes);
-
-    lib.write_trailer_into(&mut output_bytes);
-
-    (output_bytes, spoofed_digest)
-}
-
-#[test]
-fn regression_library_deserialisation_rejects_spoofed_mast_node_digests() {
-    let lib_src = r#"
-pub proc p
-    push.1
-end
-"#;
-    let lib = Assembler::default()
-        .assemble_library("lib", [lib_src])
-        .expect("library assembly must succeed");
-
-    let (bytes, _) =
-        build_library_bytes_with_spoofed_first_node_digest(&lib, "spoofed-library-digest");
-    let err = Package::read_from_bytes(&bytes)
-        .expect_err("expected library deserialization to reject inconsistent node digests");
-    assert!(
-        err.to_string().contains("invalid untrusted MAST forest"),
-        "expected untrusted-MAST validation failure, got: {err}"
-    );
-    assert!(
-        err.to_string().contains("hash mismatch for node"),
-        "expected digest mismatch failure, got: {err}"
-    );
-}
-
-#[test]
-fn unchecked_library_deserialisation_accepts_spoofed_mast_node_digests() {
-    let lib_src = r#"
-pub proc p
-    push.1
-end
-"#;
-    let lib = Assembler::default()
-        .assemble_library("lib", [lib_src])
-        .expect("library assembly must succeed");
-
-    let (bytes, spoofed_digest) =
-        build_library_bytes_with_spoofed_first_node_digest(&lib, "spoofed-library-digest");
-    let deserialized = Package::read_from_bytes_unchecked(&bytes)
-        .expect("unchecked library deserialization must accept spoofed node digests");
-
-    assert_eq!(
-        deserialized.mast_forest()[MastNodeId::new_unchecked(0)].digest(),
-        spoofed_digest
-    );
-}
-
-#[test]
-fn regression_kernel_library_deserialisation_rejects_spoofed_mast_node_digests() {
-    let kernel_src = r#"
-pub proc k1
-    push.1
-end
-"#;
-    let kernel_lib = Assembler::default()
-        .assemble_kernel("kernel", kernel_src)
-        .expect("kernel assembly must succeed");
-
-    let (bytes, _) = build_library_bytes_with_spoofed_first_node_digest(
-        kernel_lib.as_ref(),
-        "spoofed-kernel-digest",
-    );
-    let err = Package::read_from_bytes(&bytes)
-        .expect_err("expected kernel library deserialization to reject inconsistent node digests");
-    assert!(
-        err.to_string().contains("invalid untrusted MAST forest"),
-        "expected untrusted-MAST validation failure, got: {err}"
-    );
-    assert!(
-        err.to_string().contains("hash mismatch for node"),
-        "expected digest mismatch failure, got: {err}"
-    );
-}
-
-#[test]
-fn unchecked_kernel_library_deserialisation_accepts_spoofed_mast_node_digests() {
-    let kernel_src = r#"
-pub proc k1
-    push.1
-end
-"#;
-    let kernel_lib = Assembler::default()
-        .assemble_kernel("kernel", kernel_src)
-        .expect("kernel assembly must succeed");
-
-    let (bytes, spoofed_digest) = build_library_bytes_with_spoofed_first_node_digest(
-        kernel_lib.as_ref(),
-        "spoofed-kernel-digest",
-    );
-    let deserialized = Package::read_from_bytes_unchecked(&bytes)
-        .expect("unchecked kernel deserialization must accept spoofed node digests");
-
-    assert_eq!(
-        deserialized.mast_forest()[MastNodeId::new_unchecked(0)].digest(),
-        spoofed_digest
-    );
-}
-
 #[test]
 fn get_module_by_path() {
     let context = TestContext::new();
